Move the outline view into a wrapper widget
