+++ novelwriter/gui/__init__.py
@@ -25,7 +25,6 @@ from novelwriter.gui.itemdetails import GuiItemDetails
 from novelwriter.gui.mainmenu import GuiMainMenu
 from novelwriter.gui.noveltree import GuiNovelTree
 from novelwriter.gui.outline import GuiOutline
-from novelwriter.gui.outlinedetails import GuiOutlineDetails
 from novelwriter.gui.projtree import GuiProjectTree
 from novelwriter.gui.statusbar import GuiMainStatus
 from novelwriter.gui.theme import GuiTheme
@@ -40,7 +39,6 @@ __all__ = [
     "GuiMainStatus",
     "GuiNovelTree",
     "GuiOutline",
-    "GuiOutlineDetails",
     "GuiProjectTree",
     "GuiTheme",
     "GuiViewsBar",
